[+] some structure changes
