+++ pom.xml
@@ -27,6 +27,5 @@
             <version>6.8</version>
             <scope>test</scope>
         </dependency>
-
     </dependencies>
 </project>
