+++ src/test/java/basetest/SingletonChrome.java
@@ -1,4 +1,4 @@
-package tests;
+package basetest;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
@@ -6,7 +6,7 @@ import org.openqa.selenium.chrome.ChromeDriver;
 public class SingletonChrome extends WebDriverOptions {
 
     private static SingletonChrome instance = null;
-    private WebDriver driver;
+    public WebDriver driver;
 
     private SingletonChrome() {
     }
