+++ testng.xml
@@ -1,7 +1,7 @@
 <!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
 
-<suite name="Suite1" verbose="1" >
-    <test name="My tests"   >
+<suite name="BBS suite" verbose="1" >
+    <test name="Find bbs.com"   >
         <packages>
             <package name="tests" />
         </packages>
